# docs/swagger.yaml
components:
  schemas:
    api.authResponse:
      properties:
        access_token:
          description: Access token
          type: string
        expires_in:
          description: Access token 有效期,单位为秒
          type: integer
        refresh_token:
          description: Refresh token
          type: string
        user:
          $ref: '#/components/schemas/api.userResponse'
      type: object
...
    api.changePasswordRequest:
      properties:
        new_password:
          description: 新密码
          maxLength: 128
          minLength: 6
          type: string
        old_password:
          description: 当前密码
          type: string
      required:
      - new_password
      - old_password
      type: object
...
    api.loginRequest:
      properties:
        password:
          description: 密码
          type: string
        username:
          description: 用户名
          type: string
      required:
      - password
      - username
      type: object
...
    api.refreshTokenRequest:
      properties:
        refresh_token:
          type: string
      required:
      - refresh_token
      type: object
    api.registerRequest:
      properties:
        password:
          description: 密码
          maxLength: 128
          minLength: 6
          type: string
        username:
          description: 用户名
          maxLength: 32
          minLength: 3
          type: string
      required:
      - password
      - username
      type: object
...
    api.updateProfileRequest:
      properties:
        avatar_url:
          description: 头像 URL
          type: string
        bio:
          description: 个人简介
          maxLength: 200
          type: string
        username:
          description: 用户名
          maxLength: 32
          minLength: 3
          type: string
      required:
      - username
      type: object
    api.userProfileResponse:
      properties:
        avatar_url:
          type: string
        bio:
          type: string
        follower_count:
          minimum: 0
          type: integer
        following_count:
          minimum: 0
          type: integer
        id:
          type: string
        post_count:
          minimum: 0
          type: integer
        received_like_and_collect_count:
          minimum: 0
          type: integer
        username:
          type: string
        viewer_following:
          type: boolean
      required:
      - avatar_url
      - bio
      - follower_count
      - following_count
      - id
      - post_count
      - received_like_and_collect_count
      - username
      - viewer_following
      type: object
    api.userResponse:
      description: 用户信息
      properties:
        avatar_url:
          description: 头像地址
          type: string
        bio:
          description: 个人简介
          type: string
        id:
          description: 用户 ID
          type: string
        username:
          description: 用户名
          type: string
      type: object
    render.Response-api_authResponse:
      properties:
        code:
          type: integer
        data:
          $ref: '#/components/schemas/api.authResponse'
        msg:
          type: string
      type: object
...
    render.Response-api_userProfileResponse:
      properties:
        code:
          type: integer
        data:
          $ref: '#/components/schemas/api.userProfileResponse'
        msg:
          type: string
      type: object
    render.Response-api_userResponse:
      properties:
        code:
          type: integer
        data:
          $ref: '#/components/schemas/api.userResponse'
        msg:
          type: string
      type: object
    render.ResponseWithoutData:
      properties:
        code:
          type: integer
        msg:
          type: string
      type: object
    render.errorResponse:
      properties:
        code:
          type: integer
        msg:
          type: string
      type: object
  securitySchemes:
    BearerAuth:
      scheme: bearer
      type: http
...
info:
  description: 小蓝书后端接口文档
  title: 小蓝书 API
  version: "1.0"
openapi: 3.1.0
paths:
  /auth/login:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.loginRequest'
              description: 登录信息
              summary: body
        description: 登录信息
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_authResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      summary: 用户登录
      tags:
      - auth
  /auth/logout:
    post:
      responses:
        "204":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.ResponseWithoutData'
          description: No Content
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 退出登录
      tags:
      - auth
  /auth/me:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_userResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Unauthorized
      security:
      - BearerAuth: []
      summary: 获取当前登录用户
      tags:
      - auth
  /auth/password:
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.changePasswordRequest'
              description: 密码信息
              summary: body
        description: 密码信息
        required: true
      responses:
        "204":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.ResponseWithoutData'
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 修改密码
      tags:
      - auth
  /auth/refresh:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.refreshTokenRequest'
              description: 刷新凭证
              summary: body
        description: 刷新凭证
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_authResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Unauthorized
      summary: 刷新登录凭证
      tags:
      - auth
  /auth/register:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.registerRequest'
              description: 注册信息
              summary: body
        description: 注册信息
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_authResponse'
          description: OK
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      summary: 用户注册
      tags:
      - auth
...
  /me/profile:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_userProfileResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 获取当前用户资料
      tags:
      - users
...
  /users/{user_id}:
    get:
      parameters:
      - description: 用户 ID
        in: path
        name: user_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_userProfileResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Not Found
      summary: 获取用户资料
      tags:
      - users
...
  /users/profile:
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.updateProfileRequest'
              description: 更新信息
              summary: body
        description: 更新信息
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_userResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 更新用户资料
      tags:
      - users
servers:
- description: 本地开发环境
  url: http://localhost:8080/api/v1